feat: adapt instillFormTree for output properties
