+++ packages/toolkit/src/lib/use-instill-form/components/component-output/ImagesField.tsx
@@ -1,0 +1,57 @@
+import { getImageTypeFromBase64String } from "../../../../view/pipeline-builder/lib/getImageTypeFromBase64";
+import { Nullable } from "../../../type";
+import { ConnectorNodeFieldRoot, EndNodeFieldRoot } from "./FieldRoot";
+
+export type ImagesFieldProps = {
+  nodeType: "end" | "connector";
+  title: Nullable<string>;
+  images: Nullable<string>[];
+};
+
+export const ImagesField = (props: ImagesFieldProps) => {
+  const { nodeType, title, images } = props;
+
+  if (nodeType === "connector") {
+    return (
+      <ConnectorNodeFieldRoot title={title} key={`${title}-field`}>
+        <div className="flex w-full flex-wrap">
+          {images?.slice(0, 5).map((image) => {
+            if (!image) return <></>;
+
+            const imageType = getImageTypeFromBase64String(image);
+
+            return (
+              <img
+                key={`${title}-${image}-field`}
+                alt={`${title}-images-{idx}`}
+                src={`data:image/${imageType};base64,${image}`}
+                className="object-contain"
+              />
+            );
+          })}
+        </div>
+      </ConnectorNodeFieldRoot>
+    );
+  }
+
+  return (
+    <EndNodeFieldRoot title={title} key={`${title}-field`}>
+      <div className="flex w-full flex-wrap">
+        {images?.slice(0, 5).map((image) => {
+          if (!image) return <></>;
+
+          const imageType = getImageTypeFromBase64String(image);
+
+          return (
+            <img
+              key={`${title}-${image}-field`}
+              alt={`${title}-images-{idx}`}
+              src={`data:image/${imageType};base64,${image}`}
+              className="object-contain"
+            />
+          );
+        })}
+      </div>
+    </EndNodeFieldRoot>
+  );
+};
