+++ packages/toolkit/src/lib/use-instill-form/components/component-output/ObjectArrayField.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
-import { Nullable } from "../../../lib";
 import { ConnectorNodeFieldRoot, EndNodeFieldRoot } from "./FieldRoot";
+import { Nullable } from "../../../type";
 
-export type ArrayObjectFieldProps = {
+export type ObjectArrayFieldProps = {
   nodeType: "end" | "connector";
   title: Nullable<string>;
   children: React.ReactNode;
 };
 
-export const ArrayObjectField = (props: ArrayObjectFieldProps) => {
+export const ObjectArrayField = (props: ObjectArrayFieldProps) => {
   const { nodeType, title, children } = props;
 
   if (nodeType === "connector") {
