+++ packages/toolkit/src/lib/use-instill-form/components/component-output/index.ts
@@ -1,0 +1,25 @@
+import { AudioField } from "./AudioField";
+import { AudiosField } from "./AudiosField";
+import { ImageField } from "./ImageField";
+import { ImagesField } from "./ImagesField";
+import { NumberField } from "./NumberField";
+import { NumbersField } from "./NumbersField";
+import { ObjectArrayField } from "./ObjectArrayField";
+import { ObjectField } from "./ObjectField";
+import { ObjectsField } from "./ObjectsField";
+import { TextField } from "./TextField";
+import { TextsField } from "./TextsField";
+
+export const ComponentOutputFields = {
+  AudioField,
+  AudiosField,
+  ImageField,
+  ImagesField,
+  NumberField,
+  NumbersField,
+  ObjectArrayField,
+  ObjectField,
+  ObjectsField,
+  TextField,
+  TextsField,
+};
