+++ packages/toolkit/src/lib/use-instill-form/components/index.ts
@@ -1,6 +1,2 @@
-export * from "./BooleanField";
-export * from "./CredentialTextField";
-export * from "./OneOfConditionField";
-export * from "./SingleSelectField";
-export * from "./TextAreaField";
-export * from "./TextField";
+export * from "./regular";
+export * from "./component-output";
