+++ packages/toolkit/src/lib/use-instill-form/components/regular/TextAreaField.tsx
@@ -2,16 +2,15 @@ import {
   Form,
   Icons,
   ParagraphWithHTML,
-  Select,
+  Textarea,
   Tooltip,
 } from "@instill-ai/design-system";
-import { GeneralUseFormReturn } from "../../type";
+import { GeneralUseFormReturn } from "../../../type";
 
-export const SingleSelectField = ({
+export const TextAreaField = ({
   form,
   path,
   title,
-  options,
   description,
   shortDescription,
   disabled,
@@ -19,14 +18,12 @@ export const SingleSelectField = ({
   form: GeneralUseFormReturn;
   path: string;
   title: string | null;
-  options: string[];
   description?: string;
   shortDescription?: string;
   disabled?: boolean;
 }) => {
   return (
     <Form.Field
-      key={path}
       control={form.control}
       name={path}
       render={({ field }) => {
@@ -64,32 +61,23 @@ export const SingleSelectField = ({
                 </Tooltip.Provider>
               ) : null}
             </div>
-            <Select.Root
-              onValueChange={(e) => {
-                field.onChange(e);
-              }}
-              value={field.value ?? undefined}
-              disabled={disabled}
-            >
-              <Form.Control>
-                <Select.Trigger className="w-full">
-                  <Select.Value />
-                </Select.Trigger>
-              </Form.Control>
-              <Select.Content>
-                {options.map((option) => {
-                  return (
-                    <Select.Item
-                      key={option}
-                      value={option}
-                      className="my-auto text-semantic-fg-primary product-body-text-2-regular group-hover:text-semantic-bg-primary data-[highlighted]:text-semantic-bg-primary"
-                    >
-                      <p className="my-auto">{option}</p>
-                    </Select.Item>
-                  );
-                })}
-              </Select.Content>
-            </Select.Root>
+            <Form.Control>
+              <Textarea
+                {...field}
+                // At some moment the value maybe a object
+                // For example, { foo: { bar: "baz" } }}}. For foo.bar field
+                // its value is a string But for foo field its value is a object.
+                // And some time the foo field is not a object field but a string field,
+                // we need to deal with it
+                value={typeof field.value === "object" ? "" : field.value ?? ""}
+                autoComplete="off"
+                onChange={(e) => {
+                  field.onChange(e);
+                  form.trigger(path, { shouldFocus: true });
+                }}
+                disabled={disabled}
+              />
+            </Form.Control>
             <Form.Description text={shortDescription ?? null} />
             <Form.Message />
           </Form.Item>
