+++ packages/toolkit/src/lib/use-instill-form/components/regular/index.ts
@@ -1,0 +1,15 @@
+import { BooleanField } from "./BooleanField";
+import { CredentialTextField } from "./CredentialTextField";
+import { OneOfConditionField } from "./OneOfConditionField";
+import { SingleSelectField } from "./SingleSelectField";
+import { TextAreaField } from "./TextAreaField";
+import { TextField } from "./TextField";
+
+export const RegularFields = {
+  BooleanField,
+  CredentialTextField,
+  OneOfConditionField,
+  SingleSelectField,
+  TextAreaField,
+  TextField,
+};
