+++ packages/toolkit/src/lib/use-instill-form/index.ts
@@ -1,2 +1,3 @@
 export * from "./type";
+export * from "./useComponentOutputFields";
 export * from "./useInstillForm";
