+++ packages/toolkit/src/view/pipeline-builder/index.ts
@@ -1,7 +1,6 @@
 export * from "./components";
 export * from "./lib";
 export * from "./use-node-input-fields";
-export * from "./use-node-output-fields";
 export * from "./Flow";
 export * from "./FlowControl";
 export * from "./RightPanel";
