--- packages/toolkit/src/view/pipeline-builder/use-node-output-fields/ImagesField.tsx
@@ -1,57 +1,0 @@
-import { Nullable } from "../../../lib";
-import { getImageTypeFromBase64String } from "../lib/getImageTypeFromBase64";
-import { ConnectorNodeFieldRoot, EndNodeFieldRoot } from "./FieldRoot";
-
-export type ImagesFieldProps = {
-	nodeType: "end" | "connector";
-	title: Nullable<string>;
-	images: Nullable<string>[];
-};
-
-export const ImagesField = (props: ImagesFieldProps) => {
-	const { nodeType, title, images } = props;
-
-	if (nodeType === "connector") {
-		return (
-			<ConnectorNodeFieldRoot title={title} key={`${title}-field`}>
-				<div className="flex w-full flex-wrap">
-					{images?.slice(0, 5).map((image) => {
-						if (!image) return <></>;
-
-						const imageType = getImageTypeFromBase64String(image);
-
-						return (
-							<img
-								key={`${title}-${image}-field`}
-								alt={`${title}-images-{idx}`}
-								src={`data:image/${imageType};base64,${image}`}
-								className="object-contain"
-							/>
-						);
-					})}
-				</div>
-			</ConnectorNodeFieldRoot>
-		);
-	}
-
-	return (
-		<EndNodeFieldRoot title={title} key={`${title}-field`}>
-			<div className="flex w-full flex-wrap">
-				{images?.slice(0, 5).map((image) => {
-					if (!image) return <></>;
-
-					const imageType = getImageTypeFromBase64String(image);
-
-					return (
-						<img
-							key={`${title}-${image}-field`}
-							alt={`${title}-images-{idx}`}
-							src={`data:image/${imageType};base64,${image}`}
-							className="object-contain"
-						/>
-					);
-				})}
-			</div>
-		</EndNodeFieldRoot>
-	);
-};
