--- packages/toolkit/src/view/pipeline-builder/use-node-output-fields/index.ts
@@ -1,2 +1,0 @@
-export * from "./useEndOperatorTestModeOutputFields";
-export * from "./useConnectorTestModeOutputFields";
